requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioice==0.10.2
aiortc==1.14.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
av==16.1.0
babel==2.18.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
cryptography==46.0.5
csvw==3.7.0
ctranslate2==4.7.1
daily-python==0.23.0
distro==1.9.0
dlinfo==2.0.0
dnspython==2.8.0
docstring_parser==0.17.0
espeakng-loader==0.2.4
fastapi==0.133.1
faster-whisper==1.2.1
filelock==3.24.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
future==1.0.0
google-crc32c==1.8.0
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
humanfriendly==10.0
idna==3.11
ifaddr==0.2.0
isodate==0.7.2
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kokoro-onnx==0.5.0
language-tags==1.2.0
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.10.2
markdown-it-py==4.0.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mysql-connector-python==9.6.0
nltk==3.9.3
numba==0.61.2
numpy==2.2.6
onnxruntime==1.23.2
openai==2.24.0
opencv-python==4.13.0.92
packaging==26.0
phonemizer-fork==3.3.2
pillow==12.1.1
pipecat-ai==0.0.103
pipecat-ai-flows==0.0.22
propcache==0.4.1
protobuf==5.29.6
PyAudio==0.2.14
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.19.2
pylibsrtp==1.0.0
pyloudnorm==0.1.1
pyOpenSSL==25.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
rdflib==7.6.0
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
resampy==0.4.3
rfc3986==1.5.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scipy==1.17.1
segments==2.3.0
setuptools==82.0.0
shellingham==1.5.4
simli-ai==2.0.2
six==1.17.0
sniffio==1.3.1
soxr==0.5.0.post1
starlette==0.52.1
sympy==1.14.0
termcolor==3.3.0
tokenizers==0.22.2
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
yarl==1.22.0
